requirements.txt
Django==2.2.7
djangorestframework==3.11.0
drf-yasg==1.17.0